Refactorisation, modularisation et optimisation (WIP): - Remplacé plusieurs boucles while par des boucles for pour une meilleure lisibilité - Création des classes Transaction et Account pour regrouper les fonctions sous forme de méthodes - Gestion de fichier open/close remplacé par with - Séparation en modules par thèmes (transactions, comptes, statistiques, etc.) - Ajustement des imports en conséquence - Transition vers le self comme argument maintenant qu’ils sont en format objet et appel de ses méthodes en incluant la classe qui les contient.
