+++ src/piledger/data_handler.py
@@ -1,5 +1,5 @@
 import csv
-
+from piledger.accounts import Account
 def read_data_file():
     data = []
     with open('data.csv', 'r', encoding='utf-8') as file:
@@ -18,14 +18,34 @@ def read_data_file():
     return data
 
 def export_account_postings(data, account_name, filename):
-    file = open(filename, 'w', encoding='utf-8')
-    file.write("No txn,Date,Compte,Montant,Commentaire\n")
-    i = 0
-    while i < len(data):
-        transaction = data[i]
-        if transaction['compte'] == account_name:
-            line = f"{transaction['no_txn']},{transaction['date']},{transaction['compte']},{transaction['montant']},{transaction['commentaire']}\n"
-            file.write(line)
-        i += 1
-    file.close()
+    with open(filename, 'w', encoding='utf-8') as file:
+        file.write("No txn,Date,Compte,Montant,Commentaire\n")
+        for transaction in data:
+            if transaction['compte'] == account_name:
+                line = f"{transaction['no_txn']},{transaction['date']},{transaction['compte']},{transaction['montant']},{transaction['commentaire']}\n"
+                file.write(line)
     print(f"Écritures exportées vers {filename}")
+
+
+def handle_export(data, accounts):
+    print("\n--- Exportation ---")
+    print("Comptes disponibles:")
+    for account in accounts:
+        print(f"  - {account}")
+    
+    account_input = input("\nEntrez le nom du compte à exporter: ").strip()
+    
+    if not account_input:
+        print("Nom de compte invalide!")
+        return
+    ac = Account(data)
+    validated_account = ac.validate_account_name(account_input)
+    
+    if validated_account:
+        filename = input("Nom du fichier de sortie (ex: export.csv): ").strip()
+        if not filename:
+            filename = f"export_{validated_account.replace(' ', '_').lower()}.csv"
+        
+        export_account_postings(data, validated_account, filename)
+    else:
+        print(f"Compte '{account_input}' introuvable!")
