+++ src/piledger/main.py
@@ -1,165 +1,11 @@
 import os
-from piledger.data_handler import read_data_file, export_account_postings
-from piledger.transactions import display_all_transactions, display_transactions_by_account, get_transactions_by_date_range
-from piledger.accounts import calculate_balance, get_all_accounts, validate_account_name
+from piledger.data_handler import read_data_file, handle_export
+from piledger.transactions import Transaction
+from piledger.accounts import Account
+from piledger.ui import display_summary, display_menu
+from piledger.statistics import handle_balance_inquiry, handle_statistics, handle_date_search
 
 
-
-def display_summary(data):
-    print("\n=== RÉSUMÉ DES COMPTES ===")
-    accounts = get_all_accounts(data)
-    i = 0
-    while i < len(accounts):
-        account = accounts[i]
-        balance = calculate_balance(data, account)
-        print(f"{account}: {balance:.2f}$")
-        i += 1
-
-
-def find_largest_expense(data):
-    largest_expense = None
-    max_amount = 0
-    i = 0
-    while i < len(data):
-        transaction = data[i]
-        if transaction['montant'] > max_amount and transaction['compte'] != 'Compte courant' and transaction['compte'] != 'Revenu':
-            max_amount = transaction['montant']
-            largest_expense = transaction
-        i += 1
-    return largest_expense
-
-def find_total_income(data):
-    total = 0
-    i = 0
-    while i < len(data):
-        transaction = data[i]
-        if transaction['compte'] == 'Revenu':
-            total += abs(transaction['montant'])
-        i += 1
-    return total
-
-def find_total_expenses(data):
-    total = 0
-    i = 0
-    while i < len(data):
-        transaction = data[i]
-        if transaction['compte'] != 'Compte courant' and transaction['compte'] != 'Revenu' and transaction['montant'] > 0:
-            total += transaction['montant']
-        i += 1
-    return total
-
-
-
-def display_menu():
-    print("\n" + "="*50)
-    print("SYSTÈME DE GESTION COMPTABLE PERSONNEL")
-    print("="*50)
-    print("1. Afficher le solde d'un compte")
-    print("2. Afficher toutes les transactions")
-    print("3. Afficher les transactions d'un compte")
-    print("4. Afficher le résumé de tous les comptes")
-    print("5. Afficher les statistiques")
-    print("6. Exporter les écritures d'un compte")
-    print("7. Rechercher par période")
-    print("0. Quitter")
-    print("="*50)
-
-def handle_balance_inquiry(data, accounts):
-    print("\n--- Consultation de solde ---")
-    print("Comptes disponibles:")
-    i = 0
-    while i < len(accounts):
-        print(f"  - {accounts[i]}")
-        i += 1
-    
-    account_input = input("\nEntrez le nom du compte: ").strip()
-    
-    if not account_input:
-        print("Nom de compte invalide!")
-        return
-    
-    validated_account = validate_account_name(accounts, account_input)
-    
-    if validated_account:
-        balance = calculate_balance(data, validated_account)
-        print(f"\nSolde du compte '{validated_account}': {balance:.2f}$")
-        
-    else:
-        print(f"Compte '{account_input}' introuvable!")
-        print("Vérifiez l'orthographe ou choisissez un compte dans la liste.")
-
-def handle_statistics(data):
-    print("\n=== STATISTIQUES FINANCIÈRES ===")
-    
-    total_income = find_total_income(data)
-    total_expenses = find_total_expenses(data)
-    net_worth = total_income - total_expenses
-    
-    print(f"Revenus totaux: {total_income:.2f}$")
-    print(f"Dépenses totales: {total_expenses:.2f}$")
-    print(f"Situation nette: {net_worth:.2f}$")
-    
-    if net_worth > 0:
-        print("📈 Situation financière positive")
-    elif net_worth < 0:
-        print("📉 Situation financière négative")
-    else:
-        print("⚖️  Situation financière équilibrée")
-    
-    largest_expense = find_largest_expense(data)
-    if largest_expense:
-        print(f"\nPlus grosse dépense: {largest_expense['montant']:.2f}$ ({largest_expense['compte']})")
-        if largest_expense['commentaire']:
-            print(f"Commentaire: {largest_expense['commentaire']}")
-    
-    current_account_balance = calculate_balance(data, 'Compte courant')
-    print(f"\nSolde du compte courant: {current_account_balance:.2f}$")
-
-def handle_date_search(data):
-    print("\n--- Recherche par période ---")
-    start_date = input("Date de début (YYYY-MM-DD): ").strip()
-    end_date = input("Date de fin (YYYY-MM-DD): ").strip()
-    
-    if not start_date or not end_date:
-        print("Dates invalides!")
-        return
-    
-    filtered_data = get_transactions_by_date_range(data, start_date, end_date)
-    
-    if len(filtered_data) == 0:
-        print(f"Aucune transaction trouvée entre {start_date} et {end_date}")
-    else:
-        print(f"\n{len(filtered_data)} écritures(s) trouvée(s) entre {start_date} et {end_date}:")
-        i = 0
-        while i < len(filtered_data):
-            transaction = filtered_data[i]
-            print(f"  {transaction['date']} - {transaction['compte']}: {transaction['montant']:.2f}$")
-            i += 1
-
-def handle_export(data, accounts):
-    print("\n--- Exportation ---")
-    print("Comptes disponibles:")
-    i = 0
-    while i < len(accounts):
-        print(f"  - {accounts[i]}")
-        i += 1
-    
-    account_input = input("\nEntrez le nom du compte à exporter: ").strip()
-    
-    if not account_input:
-        print("Nom de compte invalide!")
-        return
-    
-    validated_account = validate_account_name(accounts, account_input)
-    
-    if validated_account:
-        filename = input("Nom du fichier de sortie (ex: export.csv): ").strip()
-        if not filename:
-            filename = f"export_{validated_account.replace(' ', '_').lower()}.csv"
-        
-        export_account_postings(data, validated_account, filename)
-    else:
-        print(f"Compte '{account_input}' introuvable!")
 
 def main():
     print("Chargement des données...")
@@ -177,7 +23,9 @@ def main():
     
     print(f"✅ {len(data)} transactions chargées avec succès!")
     
-    accounts = get_all_accounts(data)
+    t = Transaction(data)
+    ac = Account(data)
+    accounts = ac.get_all_accounts()
     
     running = True
     while running:
@@ -192,21 +40,19 @@ def main():
         if choice == "1":
             handle_balance_inquiry(data, accounts)
         elif choice == "2":
-            display_all_transactions(data)
+            t.display_all_transactions()
         elif choice == "3":
             print("\n--- Transactions par compte ---")
             print("Comptes disponibles:")
-            i = 0
-            while i < len(accounts):
-                print(f"  - {accounts[i]}")
-                i += 1
+            for account in accounts:
+                print(f"  - {account}")
             
             account_input = input("\nEntrez le nom du compte: ").strip()
             
             if account_input:
-                validated_account = validate_account_name(accounts, account_input)
+                validated_account = ac.validate_account_name(account_input)
                 if validated_account:
-                    display_transactions_by_account(data, validated_account)
+                    t.display_transactions_by_account(validated_account)
                 else:
                     print(f"Compte '{account_input}' introuvable!")
             else:
